autofs-5.1.1 - add configuration option to use fqdn in mounts
When a server name returns multiple IP addresses autofs uses the IP
address when performing the mount to ensure that the the host proximity
order is respected, and that servers that aren't responding aren't
tried.

But sometimes people need to use the server name for the mount so
add a configuration option to enable that.

Signed-off-by: Ian Kent <raven@themaw.net>