autofs-5.1.1 - fix out of order call in program map lookup
Commit 91e42e58b4 fixed a problem with program map environment variable
naming and commit 743deb0e4e added a configuration option to force use
of the old environment names for those who need it and are sure it is
safe to continue to use them.

But the call to get the configuration entry was placed after a fork()
so the state of the mutex used when querying the configuration is
undefined and can lead to a deadlock.

Signed-off-by: Ian Kent <raven@themaw.net>