autofs-5.1.1 - Add a mode option for master map entries
Add a --mode map option to change the mode for the base location mount
point. If this option is given, autofs will chmod the mount point right
after mounting it (as the kernel autofs filesystem doesn't support a
'mode' option).

Changing the mode of the base location mount point is normally not needed,
but if one wants to do that, it's much better to do it inside autofs rather
than outside to avoid race conditions and making sure the correct permissions
are always set.

Signed-off-by: Cyril B <cbay@excellency.fr>
Signed-off-by: Ian Kent <raven@themaw.net>