autofs-5.1.1 - revert fix libtirpc name clash
Commit 09d4edb2 introduced a substitution maro for the tirpc macro
auth_destroy() because it contained an unconditional call to a function
log_debug() which clashed with an autofs function of the same name.

AFAICS that was never in a actual libtirpc release, the call in a release
was auth_log_debug() which doesn't clash with any autofs functions.

So revert this change.

Signed-off-by: Ian Kent <raven@themaw.net>