autofs-5.1.1 - always set direct mounts catatonic at exit
Direct mounts are all mounted at application start or when the map
is re-read and entries have been added.

They are only ever umounted at application exit or when the map is
re-read and entries have been removed.

If these mounts are in use (so that they are not umounted) and aren't
set catatonic at exit and an application attempts to access the path
it will lead to a hang as there is no daemon to answer the mount
request.

It's questionable whether to set busy direct mounts catatonic when
attempting to umount them when re-reading the map as the mount may
then expire leaving an unresponsive direct mount trigger that hasn't
yet been cleaned from the map entry cache.

Signed-off-by: Ian Kent <raven@themaw.net>