autofs-5.1.1 - fix handle_mounts() termination condition check
In get_pkt(), if a kernel request is present on the kernel pipe and
the autofs mount point state changes to ST_SHUTDOWN after the poll(2)
check but before the request has been processed the handle_mounts()
thread will exit without shutting down the autofs mount point.

So change the handle_mounts() exit condition check to take account
of this case.

Signed-off-by: Ian Kent <raven@themaw.net>