autofs-5.1.1 - add remote-fs.target systemd dependency
If maps are stored on a NFS mount autofs will fail to start if the
systemd unit doesn't depend on the remote-fs target.

I'm not sure how this will affect systems that don't have anything
to depend on in the remote-fs target but I expect it will not stop
the startup from functioning.

Signed-off-by: Ian Kent <raven@themaw.net>